github: add scan to releases
rather than having the scan of both source and image running for _just_
PRs, have it also as part of commits that get pushed to `main`.

this way older PRs that have been accepted but not yet merged get a
second chance of having their contents scanned as they get into main.

Signed-off-by: Ciro S. Costa <ciroscosta@vmware.com>